+++ Source/SwiftLintFramework/Rules/TrailingWhitespaceRule.swift
@@ -10,9 +10,9 @@ import Foundation
 import SourceKittenFramework
 
 public struct TrailingWhitespaceRule: CorrectableRule, ConfigurationProviderRule,
-SourceKitFreeRule {
+                                      SourceKitFreeRule {
 
-    public var configuration = SeverityConfiguration(.Warning)
+    public var configuration = TrailingWhitespaceConfiguration(ignoresEmptyLines: false)
 
     public init() {}
 
@@ -26,11 +26,16 @@ SourceKitFreeRule {
     )
 
     public func validateFile(file: File) -> [StyleViolation] {
-        return file.lines.filter {
-            $0.content.hasTrailingWhitespace()
-        }.map {
+        let filteredLines = file.lines.filter {
+            $0.content.hasTrailingWhitespace() &&
+                (!configuration.ignoresEmptyLines ||
+                    // If configured, ignore lines that contain nothing but whitespace (empty lines)
+                    !$0.content.stringByTrimmingCharactersInSet(.whitespaceCharacterSet()).isEmpty)
+        }
+
+        return filteredLines.map {
             StyleViolation(ruleDescription: self.dynamicType.description,
-                severity: configuration.severity,
+                severity: configuration.severityConfiguration.severity,
                 location: Location(file: file.path, line: $0.index))
         }
     }
@@ -43,6 +48,12 @@ SourceKitFreeRule {
         for line in file.lines {
             let correctedLine = (line.content as NSString)
                 .stringByTrimmingTrailingCharactersInSet(whitespaceCharacterSet)
+
+            if configuration.ignoresEmptyLines && correctedLine.characters.isEmpty {
+                correctedLines.append(line.content)
+                continue
+            }
+
             let region = fileRegions.filter {
                 $0.contains(Location(file: file.path, line: line.index, character: 0))
             }.first
@@ -50,6 +61,7 @@ SourceKitFreeRule {
                 correctedLines.append(line.content)
                 continue
             }
+
             if line.content != correctedLine {
                 let description = self.dynamicType.description
                 let location = Location(file: file.path, line: line.index)